Move shared validation logic to shared directory
