+++ contentcuration/contentcuration/frontend/channelEdit/components/AnswersEditor/AnswersEditor.vue
@@ -104,9 +104,10 @@
 
 <script>
 
-  import { AssessmentItemTypes, AssessmentItemToolbarActions } from '../../constants';
+  import { AssessmentItemToolbarActions } from '../../constants';
   import { getCorrectAnswersIndices, mapCorrectAnswers } from '../../utils';
   import AssessmentItemToolbar from '../AssessmentItemToolbar';
+  import { AssessmentItemTypes } from 'shared/constants';
   import { swapElements } from 'shared/utils/helpers';
   import Checkbox from 'shared/views/form/Checkbox';
