+++ contentcuration/contentcuration/frontend/channelEdit/components/AssessmentItemPreview/AssessmentItemPreview.vue
@@ -108,9 +108,10 @@
 <script>
 
   import sortBy from 'lodash/sortBy';
-  import { AssessmentItemTypes, AssessmentItemTypeLabels } from '../../constants';
+  import { AssessmentItemTypeLabels } from '../../constants';
   import { getCorrectAnswersIndices } from '../../utils';
   import translator from '../../translator';
+  import { AssessmentItemTypes } from 'shared/constants';
 
   import Checkbox from 'shared/views/form/Checkbox';
   import MarkdownViewer from 'shared/views/MarkdownEditor/MarkdownViewer/MarkdownViewer.vue';
