+++ contentcuration/contentcuration/frontend/channelEdit/translator.js
@@ -16,19 +16,4 @@ const MESSAGES = {
   errorProvideAtLeastOneCorrectAnswer: 'Provide at least one correct answer',
 };
 
-const VALIDATION_MESSAGES = {
-  titleRequired: 'Title is required',
-  licenseRequired: 'License is required',
-  copyrightHolderRequired: 'Copyright holder is required',
-  licenseDescriptionRequired: 'Special permissions license must have a description',
-  masteryModelRequired: 'Mastery is required',
-  masteryModelMRequired: 'Required',
-  masteryModelMGtZero: 'Must be at least 1',
-  masteryModelMWholeNumber: 'Must be a whole number',
-  masteryModelMLteN: 'Must be lesser than or equal to N',
-  masteryModelNRequired: 'Required',
-  masteryModelNGtZero: 'Must be at least 1',
-  masteryModelNWholeNumber: 'Must be a whole number',
-};
-
-export default createTranslator(NAMESPACE, { ...MESSAGES, ...VALIDATION_MESSAGES });
+export default createTranslator(NAMESPACE, MESSAGES);
