+++ contentcuration/contentcuration/frontend/channelEdit/vuex/contentNode/getters.js
@@ -3,7 +3,7 @@ import sortBy from 'lodash/sortBy';
 import uniq from 'lodash/uniq';
 import uniqBy from 'lodash/uniqBy';
 
-import { validateNodeDetails, validateNodeFiles } from '../../utils';
+import { validateNodeDetails, validateNodeFiles } from 'shared/utils/validation';
 import { isSuccessor } from 'shared/utils/helpers';
 import { ContentKindsNames } from 'shared/leUtils/ContentKinds';
