+++ contentcuration/contentcuration/frontend/shared/constants.js
@@ -111,3 +111,31 @@ export const ErrorTypes = Object.freeze({
   // When user does not have object-level permissions (e.g. cannot edit a channel)
   UNAUTHORIZED: 'UNAUTHORIZED',
 });
+
+// should correspond to backend types
+export const AssessmentItemTypes = {
+  SINGLE_SELECTION: 'single_selection',
+  MULTIPLE_SELECTION: 'multiple_selection',
+  TRUE_FALSE: 'true_false',
+  INPUT_QUESTION: 'input_question',
+  PERSEUS_QUESTION: 'perseus_question',
+};
+
+export const ValidationErrors = {
+  TITLE_REQUIRED: 'TITLE_REQUIRED',
+  LICENSE_REQUIRED: 'LICENSE_REQUIRED',
+  COPYRIGHT_HOLDER_REQUIRED: 'COPYRIGHT_HOLDER_REQUIRED',
+  LICENSE_DESCRIPTION_REQUIRED: 'LICENSE_DESCRIPTION_REQUIRED',
+  MASTERY_MODEL_REQUIRED: 'MASTERY_MODEL_REQUIRED',
+  MASTERY_MODEL_M_REQUIRED: 'MASTERY_MODEL_M_REQUIRED',
+  MASTERY_MODEL_M_WHOLE_NUMBER: 'MASTERY_MODEL_M_WHOLE_NUMBER',
+  MASTERY_MODEL_M_GT_ZERO: 'MASTERY_MODEL_M_GT_ZERO',
+  MASTERY_MODEL_M_LTE_N: 'MASTERY_MODEL_M_LTE_N',
+  MASTERY_MODEL_N_REQUIRED: 'MASTERY_MODEL_N_REQUIRED',
+  MASTERY_MODEL_N_WHOLE_NUMBER: 'MASTERY_MODEL_N_WHOLE_NUMBER',
+  MASTERY_MODEL_N_GT_ZERO: 'MASTERY_MODEL_N_GT_ZERO',
+  QUESTION_REQUIRED: 'QUESTION_REQUIRED',
+  INVALID_NUMBER_OF_CORRECT_ANSWERS: 'INVALID_NUMBER_OF_CORRECT_ANSWERS',
+  NO_VALID_PRIMARY_FILES: 'NO_VALID_PRIMARY_FILES',
+  ...fileErrors,
+};
